extract message tracking section from dashboard page
