+++ src/app/(protected)/dashboard/dashboard-cards.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import {Mail, RefreshCw} from "lucide-react"
 import {Button} from "@/components/ui/button"
 import {Card, CardContent, CardHeader} from "@/components/ui/card"
